gcp/google-cloud-functions/requirements.txt
numpy==1.19.5
tensorflow==2.7.0
google-cloud-storage==1.16.1